+++ internal/services/tour/create.go
@@ -5,6 +5,6 @@ import (
 	"localx/internal/models"
 )
 
-func (t *TourService) CreateTour(ctx context.Context, tour models.Tour, companyId int) (int, error) {
+func (t *TourService) CreateTour(ctx context.Context, tour models.Tour, companyId int64) (int64, error) {
 	return t.repo.CreateTour(ctx, tour, companyId)
 }
